==> Analysis/01_GWAS_Prep.Rmd <==
---
title: "GWAS_Prep"
output:
  html_document:
    toc: true
    # toc_float: true
    # theme: united
    theme: cerulean
    # number_sections: true
date: "`r Sys.Date()`"
author: "Mikhail Dozmorov"
---

```{r setup, echo=FALSE, message=FALSE, warning=FALSE}
# Set up the environment
library(knitr)
opts_chunk$set(cache.path='cache/', fig.path='img/', cache=F, tidy=T, fig.keep='high', echo=F, dpi=100, warnings=F, message=F, comment=NA, warning=F, results='as.is', fig.width = 10, fig.height = 6) #out.width=700, 
library(pander)
panderOptions('table.split.table', Inf)
panderOptions('knitr.auto.asis', FALSE)
set.seed(1)
library(dplyr)
options(stringsAsFactors = FALSE)
```

```{r}
# Read in data and packages
# install.packages("Biocmanager")
library(gwascat) # BiocManager::install("gwascat")
library(limma) # BiocManager::install("limma")
library(dplyr)   # install.packages("dplyr")
source("https://raw.githubusercontent.com/mdozmorov/MDmisc/master/R/unembed.R") # library(MDmisc)  
# install.packages("rJava")
library(rowr)    # install.packages("rowr")
library(readr)   # install.packages("readr")
 
```

```{r}
#Initial summary
data(ebicat37)

#Convert to data frame
ebi_dat = as.data.frame(mcols(ebicat37))

#Put comma seperated genes on different lines
ebi_dat = unembed(ebi_dat, "MAPPED_GENE", sep = ",")
ebi_dat = unembed(ebi_dat, "MAPPED_GENE", sep = " - ")

#Remove extra whitespace from unmebed
ebi_dat$MAPPED_GENE = gsub(" ", "", ebi_dat$MAPPED_GENE)

#Summarise
init_sum = ebi_dat %>% dplyr::select(DISEASE.TRAIT, MAPPED_GENE, SNPS)  %>% distinct() %>% group_by(DISEASE.TRAIT) %>% summarise(snp_count_total = length(unique(SNPS)), snp_count_no_gene = sum(MAPPED_GENE == ""), snp_prop_nogene = snp_count_no_gene/snp_count_total,gene_count_total = length(unique(MAPPED_GENE)) - sum(unique(MAPPED_GENE =="")))
```

```{r }
#Converting aliases to gene names
alias2SymbolEbicat = function(ebicat37) {
  
  ebi_dat = as.data.frame(mcols(ebicat37))
  
  #Unembed comma and - seperated values
  ebi_dat = unembed(ebi_dat, "MAPPED_GENE", sep = ",")
  ebi_dat = unembed(ebi_dat, "MAPPED_GENE", sep = " - ")
  
  #Remove extra spaces
  ebi_dat$MAPPED_GENE = gsub(" ", "", ebi_dat$MAPPED_GENE)

  #Summarise the results
  #Keeping track of changes (sapply necessary to keep placeholders for unmapped genes)
  recomb_genes = sapply(ebi_dat$MAPPED_GENE, function(x) alias2Symbol(x, species = "Hs"))
  recomb_genes = unlist(lapply(recomb_genes, function(x) paste0(x,collapse=' ')))
  change = (ebi_dat$MAPPED_GENE == recomb_genes)
  sum_vec = ifelse(change == TRUE, "Same", ifelse(change == FALSE & recomb_genes == "", "No Match", "Converted"))
  ebi_dat$MAPPED_GENE = recomb_genes
  return(ebi_dat)
}

ebi_dat = alias2SymbolEbicat(ebicat37)

#Pulling out useful info
main_inf = ebi_dat %>% dplyr::select(MAPPED_GENE, SNPS, CHR_ID, CHR_POS, INITIAL.SAMPLE.DESCRIPTION, MAPPED_TRAIT, MAPPED_TRAIT_URI, DISEASE.TRAIT)

```

```{r}
#Collapse genes and disease names for simplifcation purposes and then get unique combinations
names_genes = main_inf %>% ungroup() %>% dplyr::select(DISEASE.TRAIT, MAPPED_GENE) %>% distinct()

#Loop through disease traits and print a gene list for each
dir.create('./data/Disease_Genes')
dis_dir = "./data/Disease_Genes/"

for (i in unique(names_genes$DISEASE.TRAIT)) {
  genes_curr = names_genes %>% filter(DISEASE.TRAIT == i & MAPPED_GENE != "") %>% dplyr::select(DISEASE.TRAIT, MAPPED_GENE) %>% mutate(DISEASE.TRAIT = gsub("\\/", "-", DISEASE.TRAIT), DISEASE.TRAIT = gsub("\\*", "-", DISEASE.TRAIT), File_Path =  DISEASE.TRAIT[1]) %>% ungroup() %>% dplyr::select(DISEASE.TRAIT, MAPPED_GENE, File_Path) %>% distinct()
  genes_disease = genes_curr %>% ungroup() %>% dplyr::select(MAPPED_GENE) %>% distinct()
  genes_curr = unique(genes_curr)
  write.table(genes_disease, paste0(dis_dir, substr(genes_curr$File_Path[1],1,100), ".txt"), row.names = FALSE, col.names = FALSE)
}
```

```{r}
# main_sum = main_inf %>% ungroup() %>%  dplyr::select(DISEASE.TRAIT, MAPPED_GENE, Category, SNPS)  %>% distinct() %>% group_by(DISEASE.TRAIT) %>% mutate(snp_count_total = length(unique(SNPS)), snp_count_no_gene = sum(MAPPED_GENE == ""), snp_prop_nogene = snp_count_no_gene/snp_count_total,gene_count_total = length(unique(MAPPED_GENE))) %>% dplyr::select(DISEASE.TRAIT, Category, snp_count_total, snp_count_no_gene, snp_prop_nogene, gene_count_total) %>% distinct()
# 
# write.csv(main_sum, "supplementary_table_diseases_eof.csv", row.names = FALSE)
```
