==> Analysis/03_Normalized_Coherence.Rmd <==
---
title: "Normalized Coherence"
author: "Kellen Cresswell"
date: "`r Sys.Date()`"
output: pdf_document
---
  
```{r setup, echo=FALSE, message=FALSE, warning=FALSE}
# Set up the environment
library(knitr)
opts_chunk$set(cache.path='cache/', fig.path='img/', cache=T, tidy=T, fig.keep='high', echo=F, dpi=100, warnings=F, message=F, comment=NA, warning=F, results='as.is', fig.width = 10, fig.height = 6) #out.width=700, 
library(pander)
panderOptions('table.split.table', Inf)
set.seed(1)
library(dplyr)
options(stringsAsFactors = FALSE)
```

# Read in functions

```{r}  
#Read in function for converting genes into data frame of internal/external edges
source("./functions/disease2frame.R") 
#Read in function for randomly sampling gene networks
source("./functions/GeneSample.R")

library(readr)   # install.packages("readr")
library(dplyr)   # install.packages("dplyr")
library(biomaRt) # BiocManager::install("biomaRt")
library(msigdbr) # BiocManager::install("msigdbr")
library(igraph)  # install.packages("igraph")
library(tidyr)   # install.packages("tidyr")
library(broom)   # install.packages("broom")
```

# Setting options

```{r}
# Option for number of msig pathways to take from each category
num_msig = 10
# Option for number of random networks to generate (Will generate num_rand-1 networks)
num_rand = 101
# Minimum number of genes having non-zero internal edges for a network to be processed
num_genes = 10
# Construct file name to save the results
run <- 1 # Placeholder for the run ID
fileNameOut <- file.path("./data/Coherence_Results/", paste0("coherence_msig_", num_msig, "_rand_", num_rand, "_", run, ".csv"))
```

# Read in Biogrid data

```{r}
# Reading in Biogrid data
biogrid = read_tsv("./data/BIOGRID-ALL-3.5.174.mitab.txt")
biogrid = biogrid %>% filter(`Taxid Interactor A` == "taxid:9606" & `Taxid Interactor B` == "taxid:9606")

#Split based on Biogrid ID system then pull out gene names
protein1 = lapply(strsplit(biogrid$`Alt IDs Interactor A`, "locuslink:"), function(x) strsplit(x[2], "\\|")[[1]][1])
protein2 = lapply(strsplit(biogrid$`Alt IDs Interactor B`, "locuslink:"), function(x) strsplit(x[2], "\\|")[[1]][1])
protein1 = unlist(protein1)
protein2 = unlist(protein2)

#Place into matrix in form that disease2frame works on
biogrid = biogrid %>% mutate(protein1 =protein1, protein2 = protein2) %>% dplyr::select(protein1, protein2)
biogrid = biogrid %>% dplyr::select(hgnc_symbol_a = protein1,hgnc_symbol_b = protein2)
```

# Read in String data and create the String filtered dataset

```{r}
# Reading in Biogrid data
string = read_table2("./data/9606.protein.links.v11.0.txt.gz")
string = string %>% mutate(protein1 = gsub("9606.", "", protein1), protein2 = gsub("9606.", "", protein2))

# Read in ensemble protein ID, convert them to gene symbols
ensembl = useEnsembl(biomart="ensembl", dataset="hsapiens_gene_ensembl", mirror = "www")
genes = biomaRt::getBM(attributes=c('ensembl_peptide_id','hgnc_symbol'), mart = ensembl)

# Join gene names with string
string = left_join(string, genes, by = c("protein1" = "ensembl_peptide_id", "protein2" = "ensembl_peptide_id"))
string = left_join(string, genes, by = c("protein2" = "ensembl_peptide_id"))
string = string %>% dplyr::rename(hgnc_symbol_a = hgnc_symbol.x)
string = string %>% dplyr::rename(hgnc_symbol_b = hgnc_symbol.y)
string_filt = string %>% filter(combined_score>=500)
```


```{r}
# [??? What this function does? Write roxygen help]

get_msig_slopes = function(sub_cats, database) {
  require(msigdbr)
  msig = msigdbr()
  
  # Subset to include msig
  msig = msig %>% filter(gs_subcat %in% sub_cats)
  
  # Create empty list
  mod_over = list()
  # Set index for list entries
  index = 1
  # Loop through each pathway
  for (j in unique(msig$gs_name)) {
    #Subset to only include current pathway
    curr_msig = msig %>% filter(gs_name == j)
    #Write to table in format for disease2frame
    write.table(curr_msig$gene_symbol, file = "temp_rea_bio.txt", quote = FALSE, row.names = FALSE, col.names = FALSE)
    #Get internal/external connectivity
    diseases_curr = diseases2frame("temp_rea_bio.txt", links = database)
    #Get disease name
    diseases_curr = diseases_curr %>% mutate(Disease = j)
    mod_over[[index]] = diseases_curr
    index =  index+1
  } 
  
  # Bind together list
  mod_over = bind_rows(mod_over)
  
  # Applying categories
  msig_cats  = msig %>% 
    dplyr::select(Disease = gs_name, Category = gs_subcat) %>% 
    distinct() #Get categories in order
  
  # Joining categories
  mod_over = left_join(mod_over, msig_cats)
  
  #Getting gene counts
  mod_over = mod_over %>% group_by(Disease) %>% filter(Internal != 0) %>% mutate(Num_Genes = n())
  
  #Getting slopes from each diseases
  mod_over = mod_over %>% group_by(Disease, Category, Num_Genes) %>% do(tidy(coef(lm(sqrt(External)~ sqrt(Internal)-1, data = .))[[1]]))
  #Rename x (Default do tidy result) to Slope
  mod_over = mod_over %>% rename(Slope = x)
  return(mod_over)
}
```

```{r}
# [??? What this function does? Write roxygen help]

get_msig_ci = function(sub_cats, database) {
  require(msigdbr)
  msig = msigdbr()
  
  # Subset to include msig
  msig = msig %>% filter(gs_subcat %in% sub_cats)
  
  # Create empty list
  mod_over = list()
  # Set index for list entries
  index = 1
  # Loop through each pathway
  for (j in unique(msig$gs_name)) {
    #Subset to only include current pathway
    curr_msig = msig %>% filter(gs_name == j)
    #Write to table in format for disease2frame
    write.table(curr_msig$gene_symbol, file = "temp_rea_bio.txt", quote = FALSE, row.names = FALSE, col.names = FALSE)
    #Get internal/external connectivity
    diseases_curr = diseases2frame("temp_rea_bio.txt", links = database)
    #Get disease name
    diseases_curr = diseases_curr %>% mutate(Disease = j)
    mod_over[[index]] = diseases_curr
    index =  index+1
  } 
  
  # Bind together list
  mod_over = bind_rows(mod_over)
  
  # Applying categories
  msig_cats  = msig %>% 
    dplyr::select(Disease = gs_name, Category = gs_subcat) %>% 
    distinct() #Get categories in order
  
  # Joining categories
  mod_over = left_join(mod_over, msig_cats)
  
  #Getting gene counts
  mod_over = mod_over %>% group_by(Disease) %>% filter(Internal != 0) %>% mutate(Num_Genes = n())
  
  #Getting slopes from each diseases
  mod_over = mod_over %>% group_by(Disease, Category, Num_Genes) %>%
    do({
      mod = lm(sqrt(External) ~ sqrt(Internal)-1, data = .)
      data.frame(Lower = confint(mod)[1], Upper = confint(mod)[2])
    }) 
  #Rename x (Default do tidy result) to Slope
  return(mod_over)
}
```

# Getting slope between internal/external edges from GOCC, KEGG and REACTOME pathways

```{r}
# Setting the three categories we want
subcats = c("CC", "CP:KEGG", "CP:REACTOME")

# Getting biogrid msig slopes [??? Do not understand, elaborate in text outside the chunk]
msig_frame_biogrid = get_msig_slopes(sub_cats = subcats, database = biogrid)

# Getting string msig slopes
msig_frame_string = get_msig_slopes(sub_cats = subcats, database = string)

# Getting string filtered msig slopes
msig_frame_string_filt = get_msig_slopes(sub_cats = subcats, database = string_filt)

# Saving the files
saveRDS(msig_frame_biogrid, "Biogrid_Msig_Slopes.rds")
saveRDS(msig_frame_string, "String_Msig_Slopes.rds")
saveRDS(msig_frame_string_filt, "String_Filt_Msig_Slopes.rds")
```


# Getting all msig pathway slopes for string (Used in Supplementary Table 2)

```{r}
#Re-read msig
msig = msigdbr()

# Define all categories
all_cats = unique(msig$gs_subcat)

#Remove empty categories
all_cats = all_cats[which(all_cats != "")]

#Get slopes
msig_frame_string_all = get_msig_slopes(sub_cats = all_cats, database = string)

#Save results
saveRDS(msig_frame_string_all, "./data/coherence_results/All_Msig_String.rds")


```

# Define function for calculating slopes and counts from internal/external edges

```{r}
# [??? What this function does? Write roxygen help]
slopes_from_edges = function(edges) {
  edges %>% filter(Internal != 0)  %>%
  group_by(Disease, Category) %>% mutate(Count = n()) %>% group_by(Disease, Category, Count) %>%
  do({
    mod = lm(sqrt(External) ~ sqrt(Internal)-1, data = .)
    data.frame(Slope= coef(mod)[1])
  }) %>% arrange(Slope) %>% filter(Count > num_genes)
}
```

The above function takes internal and external edges from each gene for each disease and calculates the total number of genes and the slope of internal vs. external edges. Genes with zero internal edges are filtered and diseases with less than `r num_genes` non-zero internal edges are filtered.

# Define function for calculating confidence intervals from edges

```{r}
ci_from_edges = function(edges) {
  edges %>% filter(Internal != 0)  %>% 
    group_by(Disease, Category) %>% mutate(Count = n()) %>% group_by(Disease, Category, Count) %>%
    do({
      mod = lm(sqrt(External) ~ sqrt(Internal)-1, data = .)
      data.frame(Lower = confint(mod)[1], Upper = confint(mod)[2])
    }) %>% arrange(Lower) %>% filter(Count > num_genes)
}
```

The above function takes internal and external edges from each gene for each disease and calculates the 95% confidence interval for coherence. Genes with zero internal edges are filtered and diseases with less than `r num_genes` non-zero internal edges are filtered.

# Get edges for diseases from Biogrid (Connectivity_Comparison.Rmd)

```{r}
# Biogrid
disease_overall_biogrid     <- readRDS("biogrid_edges_new.rds")
# Get Biogrid slopes and edge counts
dis_slopes_biogrid_count = slopes_from_edges(disease_overall_biogrid)
dis_slopes_biogrid_ci = ci_from_edges(disease_overall_biogrid)
# String
disease_overall_string     <- readRDS("string_edges_new.rds")
# Get string slopes and edge counts
dis_slopes_string_count = slopes_from_edges(disease_overall_string)
dis_slopes_string_ci = ci_from_edges(disease_overall_string)

# String Filtered
disease_overall_string_filt     <- readRDS("string_filt_edges_new.rds")
# Get string filtered slopes and edge counts
dis_slopes_string_filt_count = slopes_from_edges(disease_overall_string_filt)
dis_slopes_string_filt_ci = ci_from_edges(disease_overall_string_filt)
```

# Merging together the confidence intervals

```{r}
#Renaming names before joining and manipulating

#BIOGRID
dis_slopes_biogrid_count = dis_slopes_biogrid_count %>% ungroup() %>% dplyr::select(-Count) %>% dplyr::rename(`Biogrid Slope` = Slope)
dis_slopes_biogrid_ci = dis_slopes_biogrid_ci %>% ungroup() %>% dplyr::select(-Count) %>% dplyr::rename(`Biogrid Lower` = Lower, `Biogrid Upper` = Upper)
dis_slopes_biogrid <- full_join(dis_slopes_biogrid_count, dis_slopes_biogrid_ci, by = c("Disease", "Category"))

#STRING
dis_slopes_string_count = dis_slopes_string_count %>% dplyr::rename(`String Slope` = Slope)
dis_slopes_string_ci = dis_slopes_string_ci %>% ungroup() %>% dplyr::select(-Count) %>% dplyr::rename(`String Lower` = Lower, `String Upper` = Upper)
dis_slopes_string <- full_join(dis_slopes_string_count, dis_slopes_string_ci, by = c("Disease", "Category"))

#STRING Filtered
dis_slopes_string_filt_count = dis_slopes_string_filt_count %>% ungroup() %>% dplyr::select(-Count) %>% dplyr::rename(`String Filt Slope` = Slope)
dis_slopes_string_filt_ci = dis_slopes_string_filt_ci %>% ungroup() %>% dplyr::select(-Count) %>% dplyr::rename(`String Filt Lower` = Lower, `String Filt Upper` = Upper)
dis_slopes_string_filt <- full_join(dis_slopes_string_filt_count, dis_slopes_string_filt_ci, by = c("Disease", "Category"))

#Biogrid and STRING
ci = full_join(dis_slopes_biogrid, dis_slopes_string, by = c("Disease", "Category"))

#Merging in STRING Filtered
ci = full_join(ci, dis_slopes_string_filt, by = c("Disease", "Category"))

write.csv(ci, '../manuscript/Tables/Confidence_Intervals.csv', quote = FALSE)
```

# Define function for calculating normalized coherence and counts ignoring zero internal edges

```{r}
# [??? What this function does? Write roxygen help]
calc_norm_coherence = function(msig_frame, dis_frame, database, num_msig = 10, num_rand = 10, ci = FALSE) {
  # Splitting into list by categories 
  Count_Sum = split(msig_frame %>% dplyr::select(Disease,Category, Num_Genes) %>% distinct() , msig_frame$Category)
  coherence_tab = bind_rows()
  #Loop through each disease name and get msig normalization and random normalization factors
  for (i in unique(dis_frame$Disease)) {
    # Get disease specific genes
    dis = dis_frame %>% filter(Disease == i)
    # Get msigdf pathways with cloest number of genes, taking num_msig from each category
    close_nets = bind_rows(lapply(Count_Sum,  function(x) x %>% mutate(Diff = abs(Num_Genes-dis$Count)) %>%
      arrange(Diff) %>% head(.,num_msig)))
    # Subsetting msig to only include close pathways
    msig_sub = msig_frame %>% filter(Disease %in% close_nets$Disease)
    # Take median
    mod_slope = median(msig_sub$Slope)
    # Creating num_rand random networks
    j = 1
    slope_list = list()
    while(j < num_rand) {
    # Empty data frame for disease networks
      diseases = bind_rows()
      # Generate random networks until there as many internal edges as current diseases
      while(nrow(diseases)<dis$Count) {
        # Generate random diseases network from PPI database[??? What this function does?]
        diseases_curr = GeneSample(dis$Count, database, directed = FALSE) 
        # Save internal and external edges
        diseases_curr = data.frame(Gene = names(diseases_curr$internal_connectivity), Internal = diseases_curr$internal_connectivity, External = diseases_curr$external_connectivity)
        # Filter out zeros
        diseases_curr = diseases_curr %>% filter(Internal !=0) 
        # Bind together non-zero internal/external edges (If all zero then diseases_curr is null and nothing happens)
        diseases = bind_rows(diseases, diseases_curr)
          }
          # Save slope of current network
          slope_list[[j]] = coef(lm(I(sqrt(External)) ~ I(sqrt(Internal))-1, data = diseases))[[1]]
          # Iterate
          j = j+1
  
      }
    
      # Get median of all slopes
      rand_slope = median(unlist(slope_list))

      if (!ci) {
      # Calculate coherence
      coherence = (dis$Slope - rand_slope)/(mod_slope-rand_slope)
      
      # Place into data frame with disease, coherence, msig normalization value and random slope normalization value
      coherence = data.frame(Disease = dis$Disease, Coherence = coherence, 
                             Msig_Norm = mod_slope, Rand_Norm = rand_slope, Count = dis$Count)
      coherence_tab = bind_rows(coherence, coherence_tab)

  
  } else {
  #Lower interval for coherence
      coherence_low = (dis$Upper - rand_slope)/(mod_slope-rand_slope)
      #Upper interval for coherence
      coherence_up = (dis$Lower - rand_slope)/(mod_slope-rand_slope)
      
      coherence = data.frame(Disease = dis$Disease, Coherence_Lower = coherence_low, Coherence_Upper = coherence_up,Msig_Norm = mod_slope, Rand_Norm = rand_slope, Count = dis$Count)  
      
      coherence_tab = bind_rows(coherence, coherence_tab)

  }
}
return(coherence_tab)

}
```

# Calculating normalized coherence

```{r}
# Calculating normalized coherence 
norm_coh_biogrid = calc_norm_coherence(msig_frame_biogrid, dis_slopes_biogrid_count, database = biogrid, num_msig = num_msig, num_rand = num_rand)
norm_coh_string = calc_norm_coherence(msig_frame_string, dis_slopes_string_count, database = string, num_msig = num_msig, num_rand = num_rand)
norm_coh_string_filt = calc_norm_coherence(msig_frame_string_filt, dis_slopes_string_filt_count, database = string_filt, num_msig = num_msig, num_rand = num_rand)

#Saving to .rds files
saveRDS(norm_coh_biogrid, "./data/Coherence_Results/coherence_biogrid.rds")
saveRDS(norm_coh_string, "./data/Coherence_Results/coherence_string.rds")
saveRDS(norm_coh_string_filt, "./data/Coherence_Results/coherence_string_filt.rds")
```

# Calculating CIs for normalized coherence

```{r}
# Calculating normalized coherence 
norm_coh_biogrid_ci = calc_norm_coherence(msig_frame_biogrid, dis_slopes_biogrid_ci, database = biogrid, num_msig = num_msig, num_rand = num_rand, ci = TRUE)
norm_coh_string_ci = calc_norm_coherence(msig_frame_string, dis_slopes_string_ci, database = string, num_msig = num_msig, num_rand = num_rand, ci = TRUE)
norm_coh_string_filt = calc_norm_coherence(msig_frame_string_filt, dis_slopes_string_filt_ci, database = string_filt, num_msig = num_msig, num_rand = num_rand, ci = TRUE)

#Saving to .rds files
saveRDS(norm_coh_biogrid, "./data/Coherence_Results/coherence_biogrid_ci.rds")
saveRDS(norm_coh_string, "./data/Coherence_Results/coherence_string_ci.rds")
saveRDS(norm_coh_string_filt, "./data/Coherence_Results/coherence_string_filt_ci.rds")
```


```{r eval=FALSE}
# Write general function for saving runs to .csv

save_norm_coh = function(coherence_tab, num_msig, num_rand, run, database) {
  require(dplyr)
  fileNameOut <- file.path("./data/Coherence_Results/", paste0("coherence_msig_",database,"_",num_msig, "_rand_", num_rand, "_", run, ".csv"))
  write_csv(coherence_tab, fileNameOut)
}

# Save the runs in .csv form
#Saving biogrid results to csv
save_norm_coh(norm_coh_biogrid, num_msig = num_msig, num_rand = num_rand, run = run, database = "Biogrid")

#Saving string results to csv
save_norm_coh(norm_coh_string, num_msig = num_msig, num_rand = num_rand, run = run, database = "String")

#Saving string filtered results to csv
save_norm_coh(norm_coh_string_filt, num_msig = num_msig, num_rand = num_rand, run = run, database = "String_Filtered")
```

# Generating random networks for comparison to msigdf (Category_Tables)

```{r}
#Function for generating random networks from a set of sizes
#size_vec is vector of sizes

generate_rand_slopes = function(size_vec, database) {
  j=1
  slope_list = c()
  for (i in size_vec) {
    # Empty data frame for disease networks
      diseases = bind_rows()
      # Generate random networks until there as many internal edges as current diseases
      while(nrow(diseases)<i) {
        # Generate random diseases network from PPI database[??? What this function does?]
        diseases_curr = GeneSample(i, database, directed = FALSE) 
        # Save internal and external edges
        diseases_curr = data.frame(Gene = names(diseases_curr$internal_connectivity), Internal = diseases_curr$internal_connectivity, External = diseases_curr$external_connectivity)
        # Filter out zeros
        diseases_curr = diseases_curr %>% filter(Internal !=0) 
        # Bind together non-zero internal/external edges (If all zero then diseases_curr is null and nothing happens)
        diseases = bind_rows(diseases, diseases_curr)
      }
          # Save slope of current network
          slope_list= c(slope_list,coef(lm(I(sqrt(External)) ~ I(sqrt(Internal))-1, data = diseases))[[1]])
          # Iterate
          j = j+1
    
  }
  data.frame(Count = size_vec, Slope = slope_list)
}
```

```{r}
#Read in data if you only want to run this section

# msig_frame_string = readRDS("String_Msig_Slopes.rds")

#Get msig counts (Comment out if using pre-calculated msig slopes)

msig_counts = unique((msig_frame_string %>% filter(Category == "CP:KEGG"))$Num_Genes)

#Generate the random slopes equal to msig lengths
string_random = generate_rand_slopes(msig_counts, string)

saveRDS(string_random, "./data/Coherence_Results/string_random_edges.rds")

# Small cleanup
unlink("temp_rea_bio.txt")
```
