==> Analysis/04_Coherence_Calculations.Rmd <==
---
title: "New_Data_Coherence"
author: "Kellen Cresswell"
date: "July 1, 2019"
output: pdf_document
---

```{r}
library(dplyr)   # install.packages("dplyr")
library(readr)   # install.packages("readr")
options(stringsAsFactors = FALSE)
```

```{r}
# Minimum number of genes having non-zero internal edges for a network to be processed
num_genes = 10
```

```{r}
#Read in edges generated from Connectivity_Comparison_*.Rmd
disease_overall_string      <- readRDS("string_edges_new.rds")
disease_overall_string_filt <- readRDS("string_filt_edges_new.rds")
disease_overall_biogrid     <- readRDS("biogrid_edges_new.rds")

#Calculate slopes for all diseases under each category

#STRING
dis_slopes_string_count = disease_overall_string %>% filter(Internal != 0) %>%
    group_by(Disease, Category) %>% mutate(Count = n()) %>% group_by(Disease, Category, Count) %>%
    do({
      mod = lm(sqrt(External) ~ sqrt(Internal)-1, data = .)
      data.frame(Slope= coef(mod)[1])
    }) %>% arrange(Slope) %>% filter(Count > num_genes)

#STRING Filtered
dis_slopes_string_filt_count = disease_overall_string_filt %>% filter(Internal != 0) %>%
    group_by(Disease, Category) %>% mutate(Count = n()) %>% group_by(Disease, Category, Count) %>%
    do({
      mod = lm(sqrt(External) ~ sqrt(Internal)-1, data = .)
      data.frame(Slope= coef(mod)[1])
    }) %>% arrange(Slope) %>% filter(Count > num_genes)

#BIOGRID
dis_slopes_biogrid_count = disease_overall_biogrid %>% filter(Internal != 0)  %>%
    group_by(Disease, Category) %>% mutate(Count = n()) %>% group_by(Disease, Category, Count) %>%
    do({
      mod = lm(sqrt(External) ~ sqrt(Internal)-1, data = .)
      data.frame(Slope= coef(mod)[1])
    }) %>% arrange(Slope) %>% filter(Count > num_genes)

```

```{r}
#Read in normalized coherence (Generated from Msigdf_*.R)

#KEGG

coherence_string_filt = readRDS("./data/Coherence_Results/coherence_string_filt.rds") %>% dplyr::select(Disease, Coherence)
coherence_string = readRDS("./data/Coherence_Results/coherence_string.rds") %>% dplyr::select(Disease, Coherence)
coherence_biogrid = readRDS("./data/Coherence_Results/coherence_biogrid.rds") %>% dplyr::select(Disease, Coherence)

#Full join KEGG coherence results together based on disease
coherence_disease = full_join(coherence_biogrid, coherence_string, by = "Disease")

coherence_disease = full_join(coherence_disease, coherence_string_filt, by = "Disease")

#Give columns proper name

colnames(coherence_disease) = c("Disease", "Biogrid_Coherence", "String_Coherence", "String_Filt_Coherence")

```

```{r}
#Get categories from slope frame

Cat = bind_rows(dis_slopes_string_count %>% ungroup() %>% dplyr::select(Disease, Category), dis_slopes_biogrid_count %>% ungroup() %>% dplyr::select(Disease, Category) ) %>% distinct()

#Join to all diseases frame

coherence_disease =left_join(coherence_disease, Cat, "Disease")


#Remove underscores from column names
colnames(coherence_disease) = gsub("\\_", " ", colnames(coherence_disease))

#Adding gene counts

disease_select = read_csv("./data/supplementary_table_diseases_selected.csv") %>% dplyr::select(Disease =`Disease name`, `Number of SNPs`, `Number of genes`)

#Join gene counts

coherence_disease = left_join(coherence_disease, disease_select)

coherence_disease = coherence_disease %>% dplyr::select(Disease, Category, `Number of SNPs`, `Number of genes`, `Biogrid Coherence`:`String Filt Coherence`)
```

```{r}

#Adding p-values from permutation test
#Biogrid
biogrid_p = readRDS("./data/Permutation_Results/Complete_10k_Biogrid.rds") %>% as.data.frame(.) %>% dplyr::select(-new_coherence)

#String
string_p = readRDS("./data/Permutation_Results/Complete_10k_String.rds") %>% as.data.frame(.)  %>% dplyr::select(-new_coherence)
string_p <- string_p[!duplicated(string_p[, 1]), ] # Fix "Attention deficit hyperactivity disorder" that was run twice

#String Filtered
string_filt_p = readRDS("./data/Permutation_Results/Complete_10k_String_Filt.rds") %>% as.data.frame(.)  %>% dplyr::select(-new_coherence)

#Give appropriate column names
colnames(biogrid_p) = c("Disease", "Biogrid P-value", "Ave")
colnames(string_p) = c("Disease", "String P-value", "Ave")
colnames(string_filt_p) = c("Disease", "String Filt P-value", "Ave") 

#Joining ignoring the average coherence column I saved

#Complete Disease dataset by adding p-values

coherence_disease = left_join(coherence_disease,biogrid_p %>% dplyr::select(-Ave), by = c("Disease" = "Disease"))

coherence_disease = left_join(coherence_disease,string_p %>% dplyr::select(-Ave), by = c("Disease" = "Disease"))

coherence_disease = left_join(coherence_disease,string_filt_p  %>% dplyr::select(-Ave), by = c("Disease" = "Disease"))

#Rounding data for table
coherence_disease = coherence_disease %>% mutate_if(is.numeric, round, digits = 2)

coherence_disease = coherence_disease %>% mutate(`Biogrid P-value` = round( as.numeric(as.character(`Biogrid P-value`)),4)) %>% mutate(`String P-value` = round( as.numeric(as.character(`String P-value`)),4)) %>% mutate(`String Filt P-value` = round( as.numeric(as.character(`String Filt P-value`)),4))

#Arranging by disease and category
coherence_disease =coherence_disease %>% arrange(`Category`, `Disease`)

#Getting non-zero counts for diseases
count_frame = full_join(dis_slopes_biogrid_count %>% ungroup() %>% dplyr::select(Disease, Count), dis_slopes_string_count %>% ungroup() %>% dplyr::select(Disease, Count), "Disease")

count_frame = full_join(count_frame, dis_slopes_string_filt_count %>% ungroup() %>% dplyr::select(Disease, Count), "Disease") %>% arrange(Disease)
```

```{r}
#Adding counts to full frame and reordering

coherence_disease = coherence_disease %>% arrange(Disease) %>% mutate(`Size of Biogrid Network` = count_frame$Count.x, `Size of String Network` = count_frame$Count.y,`Size of String Filt Network`=count_frame$Count) %>% dplyr::select(Disease:`Number of genes`, `Size of Biogrid Network`:`Size of String Filt Network`, `Biogrid Coherence`:`String Filt P-value`)

#Filter out extra p-values

coherence_disease = coherence_disease %>% mutate(`Biogrid P-value` = ifelse(is.na(`Biogrid Coherence`), NA, `Biogrid P-value` ))  %>% mutate(`String P-value` = ifelse(is.na(`String Coherence`), NA, `String P-value` ))  %>% mutate(`String Filt P-value` = ifelse(is.na(`String Filt Coherence`), NA, `String Filt P-value` ))  

#Writing to a tables

dir.create("../manuscript/tables")
write.csv(coherence_disease, ".././manuscript/Tables/supplementary_table_S3.csv", 
          row.names = FALSE)

```

```{r}
```
