==> Analysis/05_Category_Tables.Rmd <==
---
title: "Category_Tables"
author: "Kellen Cresswell"
date: "July 19, 2019"
output: pdf_document
---

```{r}
library(dplyr) #install.packages("dplyr")
```

```{r}
#Creating the category table

Disease_Sum = read.csv("../manuscript/tables/supplementary_table_S3.csv")

#Summarizing Categories

Disease_String_Sum = Disease_Sum %>% dplyr::select(Disease, Category, Number.of.SNPs, Number.of.genes, String.Coherence) %>% filter(!is.na(String.Coherence)) %>% group_by(Category) %>% summarise(Min_Coherence = min(String.Coherence, na.rm = TRUE), Max_Coherence = max(String.Coherence, na.rm = TRUE), Mean_Coherence = mean(String.Coherence, na.rm = TRUE), Median_Coherence =median(String.Coherence, na.rm = TRUE), Min_Genes = min(Number.of.genes, na.rm = TRUE), Max_Genes = max(Number.of.genes, na.rm = TRUE), Mean_Genes = mean(Number.of.genes, na.rm = TRUE), Median_Genes = median(Number.of.genes, na.rm = TRUE) , Min_SNPs = min(Number.of.SNPs, na.rm = TRUE), Max_SNPs = max(Number.of.SNPs, na.rm = TRUE), Mean_SNPs = mean(Number.of.SNPs, na.rm = TRUE), Median_SNPs = median(Number.of.SNPs, na.rm = TRUE), Total_Networks = n()) %>% arrange(-Median_Coherence)

Disease_String_Filt_Sum = Disease_Sum %>% dplyr::select(Disease, Category, Number.of.SNPs, Number.of.genes, String.Filt.Coherence) %>% filter(!is.na(String.Filt.Coherence))   %>% group_by(Category) %>% summarise(Min_Coherence = min(String.Filt.Coherence, na.rm = TRUE), Max_Coherence = max(String.Filt.Coherence, na.rm = TRUE), Mean_Coherence = mean(String.Filt.Coherence, na.rm = TRUE), Median_Coherence =median(String.Filt.Coherence, na.rm = TRUE), Min_Genes = min(Number.of.genes, na.rm = TRUE), Max_Genes = max(Number.of.genes, na.rm = TRUE), Mean_Genes = mean(Number.of.genes, na.rm = TRUE), Median_Genes = median(Number.of.genes, na.rm = TRUE) , Min_SNPs = min(Number.of.SNPs, na.rm = TRUE), Max_SNPs = max(Number.of.SNPs, na.rm = TRUE), Mean_SNPs = mean(Number.of.SNPs, na.rm = TRUE), Median_SNPs = median(Number.of.SNPs, na.rm = TRUE), Total_Networks = n()) %>% arrange(-Median_Coherence)

Disease_Biogrid_Sum = Disease_Sum %>% dplyr::select(Disease, Category, Number.of.SNPs, Number.of.genes, Biogrid.Coherence) %>% filter(!is.na(Biogrid.Coherence)) %>% group_by(Category) %>% summarise(Min_Coherence = min(Biogrid.Coherence, na.rm = TRUE), Max_Coherence = max(Biogrid.Coherence, na.rm = TRUE), Mean_Coherence = mean(Biogrid.Coherence, na.rm = TRUE), Median_Coherence =median(Biogrid.Coherence, na.rm = TRUE), Min_Genes = min(Number.of.genes, na.rm = TRUE), Max_Genes = max(Number.of.genes, na.rm = TRUE), Mean_Genes = mean(Number.of.genes, na.rm = TRUE), Median_Genes = median(Number.of.genes, na.rm = TRUE) , Min_SNPs = min(Number.of.SNPs, na.rm = TRUE), Max_SNPs = max(Number.of.SNPs, na.rm = TRUE), Mean_SNPs = mean(Number.of.SNPs, na.rm = TRUE), Median_SNPs = median(Number.of.SNPs, na.rm = TRUE), Total_Networks = n()) %>% arrange(-Median_Coherence)

#Formatting

#String

Disease_String_Sum = Disease_String_Sum %>% rename(`Minimum Coherence` = Min_Coherence, `Maximum Coherence` = Max_Coherence, `Mean Coherence` = Mean_Coherence, `Median Coherence` =  Median_Coherence, `Minimum # Genes` = Min_Genes, `Maximum # Genes` = Max_Genes, `Mean # Genes` = Mean_Genes, `Median # Genes` = Median_Genes, `Minimum # Genes` = Min_Genes, `Maximum # Genes` = Max_Genes, `Mean # Genes` = Mean_Genes, `Median # Genes` = Median_Genes, `Minimum # SNPs` = Min_SNPs, `Maximum # SNPs` = Max_SNPs, `Mean # SNPs` = Mean_SNPs, `Median # SNPs` = Median_SNPs, `Minimum # SNPs` = Min_SNPs, `Maximum # SNPs` = Max_SNPs, `Mean # SNPs` = Mean_SNPs, `Median # SNPs` = Median_SNPs, `Total Networks` = Total_Networks) 

Disease_String_Sum = Disease_String_Sum %>% mutate(`Minimum Coherence` = format(`Minimum Coherence`, digits = 3)) %>% mutate(`Maximum Coherence` = format(`Maximum Coherence`, digits = 3)) %>% mutate(`Mean Coherence` = format(`Mean Coherence`, digits = 3)) %>% mutate(`Median Coherence` = format(`Median Coherence`, digits = 3)) %>% mutate(`Mean # Genes` = format(`Mean # Genes`, digits = 4)) %>% mutate(`Mean # SNPs` = format(`Mean # SNPs`, digits = 4))

#String Filtered

Disease_String_Filt_Sum = Disease_String_Filt_Sum %>% rename(`Minimum Coherence` = Min_Coherence, `Maximum Coherence` = Max_Coherence, `Mean Coherence` = Mean_Coherence, `Median Coherence` =  Median_Coherence, `Minimum # Genes` = Min_Genes, `Maximum # Genes` = Max_Genes, `Mean # Genes` = Mean_Genes, `Median # Genes` = Median_Genes, `Minimum # Genes` = Min_Genes, `Maximum # Genes` = Max_Genes, `Mean # Genes` = Mean_Genes, `Median # Genes` = Median_Genes, `Minimum # SNPs` = Min_SNPs, `Maximum # SNPs` = Max_SNPs, `Mean # SNPs` = Mean_SNPs, `Median # SNPs` = Median_SNPs, `Minimum # SNPs` = Min_SNPs, `Maximum # SNPs` = Max_SNPs, `Mean # SNPs` = Mean_SNPs, `Median # SNPs` = Median_SNPs, `Total Networks` = Total_Networks) 

Disease_String_Filt_Sum = Disease_String_Filt_Sum %>% mutate(`Minimum Coherence` = format(`Minimum Coherence`, digits = 3)) %>% mutate(`Maximum Coherence` = format(`Maximum Coherence`, digits = 3)) %>% mutate(`Mean Coherence` = format(`Mean Coherence`, digits = 3)) %>% mutate(`Median Coherence` = format(`Median Coherence`, digits = 3)) %>% mutate(`Mean # Genes` = format(`Mean # Genes`, digits = 4)) %>% mutate(`Mean # SNPs` = format(`Mean # SNPs`, digits = 4))

#Biogrid

Disease_Biogrid_Sum = Disease_Biogrid_Sum %>% rename(`Minimum Coherence` = Min_Coherence, `Maximum Coherence` = Max_Coherence, `Mean Coherence` = Mean_Coherence, `Median Coherence` =  Median_Coherence, `Minimum # Genes` = Min_Genes, `Maximum # Genes` = Max_Genes, `Mean # Genes` = Mean_Genes, `Median # Genes` = Median_Genes, `Minimum # Genes` = Min_Genes, `Maximum # Genes` = Max_Genes, `Mean # Genes` = Mean_Genes, `Median # Genes` = Median_Genes, `Minimum # SNPs` = Min_SNPs, `Maximum # SNPs` = Max_SNPs, `Mean # SNPs` = Mean_SNPs, `Median # SNPs` = Median_SNPs, `Minimum # SNPs` = Min_SNPs, `Maximum # SNPs` = Max_SNPs, `Mean # SNPs` = Mean_SNPs, `Median # SNPs` = Median_SNPs, `Total Networks` = Total_Networks) 

Disease_Biogrid_Sum = Disease_Biogrid_Sum %>% mutate(`Minimum Coherence` = format(`Minimum Coherence`, digits = 3)) %>% mutate(`Maximum Coherence` = format(`Maximum Coherence`, digits = 3)) %>% mutate(`Mean Coherence` = format(`Mean Coherence`, digits = 3)) %>% mutate(`Median Coherence` = format(`Median Coherence`, digits = 3)) %>% mutate(`Mean # Genes` = format(`Mean # Genes`, digits = 4)) %>% mutate(`Mean # SNPs` = format(`Mean # SNPs`, digits = 4))

#Writing to workbook

row.names(Disease_String_Filt_Sum) = NULL
row.names(Disease_String_Sum) = NULL
row.names(Disease_Biogrid_Sum) = NULL

require(writexl)

wb = list("STRING" = Disease_String_Sum %>% arrange(Category), "STRING Filtered" = Disease_String_Filt_Sum  %>% arrange(Category), "Biogrid" = Disease_Biogrid_Sum  %>% arrange(Category))

write_xlsx(wb, "../manuscript/Tables/supplementary_table_S1.xlsx")

```

```{r}
#In text version

Cat_Within = Disease_Sum %>% group_by(Category) %>% summarise(`Number of SNPs`= sum(`Number.of.SNPs`), `Number of Genes` = sum(Number.of.genes), `Total Diseases` = n(),  `Mean Biogrid Coherence` =mean(Biogrid.Coherence, na.rm = TRUE), `Mean String Coherence` = mean(String.Coherence, na.rm = TRUE), `Mean String Filtered Coherence` = mean(String.Filt.Coherence, na.rm = TRUE)) %>% arrange(Category)

write.csv(Cat_Within, "../manuscript/Tables/table_1.csv", row.names = FALSE)
```

```{r}
msig = readRDS("./data/Coherence_Results/All_Msig_String.rds")

#Correlation betweeen categories and size

msig_cor = msig %>% group_by(Category) %>% summarise(Slope_Cor = cor(Num_Genes, Slope, "complete")) %>% arrange(Slope_Cor)

colnames(msig_cor)[2] = "Size Correlation"

```

```{r}
options(scipen = 2)

msig_sum = msig %>% group_by(Category) %>% summarise(Min_Slope = min(Slope, na.rm = TRUE), Max_Slope = max(Slope, na.rm = TRUE), Mean_Slope = mean(Slope, na.rm = TRUE), Median_Slope = median(Slope, na.rm = TRUE), Min_Genes = min(Num_Genes), Max_Genes = max(Num_Genes), Mean_Genes = mean(Num_Genes), Median_Genes = median(Num_Genes), Total_Sets = n()) %>% arrange(Mean_Slope)

msig_sum = left_join(msig_sum, msig_cor)

#Adding in random slope

rand_string = readRDS("./data/Coherence_Results/string_random_edges.rds") %>% mutate(Category = "Random") %>% dplyr::select(Category, Count, Slope)

rand_sum = rand_string %>% group_by(Category) %>% summarise(Min_Slope = min(Slope, na.rm = TRUE), Max_Slope = max(Slope, na.rm = TRUE), Mean_Slope = mean(Slope, na.rm = TRUE), Median_Slope = median(Slope, na.rm = TRUE), Min_Genes = min(Count), Max_Genes = max(Count), Mean_Genes = mean(Count), Median_Genes = median(Count), Total_Sets = n(),) %>% arrange(Mean_Slope)

rand_sum = rand_sum %>% mutate(`Size Correlation` = cor(rand_string$Count, rand_string$Slope))

msig_sum = bind_rows(msig_sum, rand_sum)

msig_sum = msig_sum %>% rename(`Minimum Slope` = Min_Slope, `Maximum Slope` = Max_Slope, `Mean Slope` = Mean_Slope, `Median Slope` =  Median_Slope, `Minimum # Genes` = Min_Genes, `Maximum # Genes` = Max_Genes, `Mean # Genes` = Mean_Genes, `Median # Genes` = Median_Genes, `Total Networks` = Total_Sets) 

msig_sum = msig_sum %>% mutate(`Minimum Slope` = format(`Minimum Slope`, digits = 3)) %>% mutate(`Maximum Slope` = format(`Maximum Slope`, digits = 3)) %>% mutate(`Mean Slope` = format(`Mean Slope`, digits = 3)) %>% mutate(`Median Slope` = format(`Median Slope`, digits = 3)) %>% mutate(`Mean # Genes` = format(`Mean # Genes`, digits = 4)) %>% mutate(`Size Correlation` = format(`Size Correlation`, digits = 2))

write_xlsx(msig_sum,"../manuscript/Tables/supplementary_table_S2.xlsx")

```

```{r}
#Creating plot of slopes from tables

#Read in counts

kegg_string  = msig %>% filter(Category == "CP:KEGG") %>% dplyr::select(Num_Genes, Slope)

Slope_Sum = bind_rows(rand_string %>% mutate(Method = "Random"), kegg_string %>% dplyr::rename(Count=Num_Genes) %>% dplyr::select(Count, Slope) %>% mutate(Method = "KEGG"))
```

```{r}
require(ggplot2)

#Plotting

slope_plot = ggplot(Slope_Sum, aes(x = Count, y = Slope, color = Method)) + geom_point(size = 2) + theme_bw() + labs(y = "Slope", x = "Network Size") + geom_smooth( method = 'loess', se = FALSE, formula=y~x-1, fullrange = TRUE, size = 2)  + theme_light(base_size = 18) + 
   theme(axis.line = element_line(colour = "black"), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        panel.background = element_blank(), legend.position = c(.75, .75))

ggsave("../manuscript/Figures/Fig2_Size_Slope_Plot.png", slope_plot, height=5, width=7, units="in", dpi = 600)
```


```{}
