==> Analysis/06_Analysis_Tables.Rmd <==
---
title: "Secondary analysis of tables"
# author: "Author's Name"
# date: "`r Sys.Date()`"
output:
  pdf_document:
    toc: no
  html_document:
    theme: cerulean
    toc: yes
editor_options: 
  chunk_output_type: console
---

```{r setup, echo=FALSE, message=FALSE, warning=FALSE}
# Set up the environment
library(knitr)
opts_chunk$set(cache.path='cache/', fig.path='img/', cache=F, tidy=T, fig.keep='high', echo=F, dpi=100, warnings=F, message=F, comment=NA, warning=F, results='as.is', fig.width = 10, fig.height = 6) #out.width=700, 
library(pander)
panderOptions('table.split.table', Inf)
set.seed(1)
library(dplyr)
options(stringsAsFactors = FALSE)
```

```{r}
library(readr)
library(ggplot2)
library(ggrepel)
library(gridExtra)
source("https://raw.githubusercontent.com/mdozmorov/MDmisc/master/R/venn.R")
library(corrplot)
col3 <- colorRampPalette(c("blue", "white", "red"))
library(openxlsx)
library(cowplot)
```

```{r}
max_genes <- 200 # Maximum number of genes for a network to be included
min_genes <- 10  # Minimum number of genes for a network to be included
fileNameIn0  <- "../manuscript/Tables/supplementary_table_S3.csv"   # diseases and traits, coherence new KEGG, REACTOME, CC. 2019-07-14, Kellen
fileNameIn1 <- "../manuscript/Tables/Confidence_Intervals.csv" # Confidence intervals, generated in 03_Normalized_Coherence 2020-03-17
# iconv -f UTF-8 -t ISO-8859-1 -c supplementary_table_S3.csv > tmp.csv
mtx0 <- read_csv(fileNameIn0) %>% as.data.frame()
# Rename some diseases
mtx0$Disease[mtx0$Disease == "Attention deficit hyperactivity disorder"] <- "ADHD"
mtx0$Disease[mtx0$Disease == "Inflammatory bowel disease"] <- "IBD"
mtx0$Disease[mtx0$Disease == "Systemic lupus erythematosus"] <- "SLE"
mtx0$Disease[mtx0$Disease == "Age-related macular degeneration"] <- "Macular degeneration"
mtx0$Disease[mtx0$Disease == "Lupus nephritis in systemic lupus erythematosus"] <- "Lupus nephritis in SLE"

rownames(mtx0) <- mtx0$Disease
```

### Append confidence intervals

```{r}
# Should be done one time
mtx <- read_csv(fileNameIn1)
mtx0 <- left_join(mtx0, mtx, by = c("Disease", "Category"))
write_csv(mtx0, fileNameIn0)
```

### Coherence distribution

```{r}
mtx0[, "String Filt Coherence"] %>% unlist %>% c %>% unname %>% hist
mtx0[, "String Filt Coherence"] %>% unlist %>% c %>% unname %>% summary
```


### FigS1_corr_coherence.png

**Supplementary Figure S1.[FILE: Figures/FigS1_corr_coherence.png] Normalized coherence is unaffected by network size.** Red/blue gradient and numbers represent Pearson correlation coefficients between corresponding entries. 

```{r}
# Add CI size
mtx0 <- mtx0 %>% mutate(`CI Biogrid`=`Biogrid Upper`-`Biogrid Lower`, `CI String`=`String Upper`-`String Lower`, `CI String Filt`=`String Filt Upper`-`String Filt Lower`)
# Selected data for correlation
# Remove small/large networks
mtx0_cor <- mtx0[ mtx0 %>% dplyr::select(starts_with("Size")) %>% apply(., 1, min, na.rm = TRUE) > min_genes &
                  mtx0 %>% dplyr::select(starts_with("Size")) %>% apply(., 1, max, na.rm = TRUE) < max_genes, ]
# Remove size columns
mtx0_cor <- mtx0_cor %>% dplyr::select(contains("Coherence"), contains("Number"), contains("P value"), contains("Modularity"), contains("Slope"), contains("CI"))
# Jitter STRING p-value a bit
set.seed(1)
mtx0_cor$`String P value` <- jitter(mtx0_cor$`String P value`, amount = 0.0001)
# Actual correlation
diff_cors <- cor(mtx0_cor, use = "pairwise.complete.obs")
# Average correlation with network size
diff_cors_selected <- diff_cors[grepl("Number", rownames(diff_cors)), grepl("Coherence", colnames(diff_cors))]
print(paste0("Average correlation with network size: ", mean(diff_cors_selected[upper.tri(diff_cors_selected)])))
# Average correlation of STRING and STRING filtered coherences
diff_cors_selected <- diff_cors[grepl("Coherence", rownames(diff_cors)) & grepl("String", rownames(diff_cors)), grepl("Coherence", colnames(diff_cors)) & grepl("String", colnames(diff_cors))]
print(paste0("Average correlation of STRING coherences: ", mean(diff_cors_selected[upper.tri(diff_cors_selected)])))
# Average correlation of coherences
diff_cors_selected <- diff_cors[grepl("Coherence", rownames(diff_cors)), grepl("Coherence", colnames(diff_cors))]
print(paste0("Average correlation of coherences: ", mean(diff_cors_selected[upper.tri(diff_cors_selected)])))
# Average correlation of coherences and p-value
diff_cors_selected <- diff_cors[grepl("Coherence", rownames(diff_cors)), grepl("P value", colnames(diff_cors))]
print(paste0("Average correlation of coherences and p-value: ", mean(diff_cors_selected[upper.tri(diff_cors_selected)])))
# Average correlation of slopes and network sizes
diff_cors_selected <- diff_cors[grepl("Slope", rownames(diff_cors)), grepl("Number of genes", colnames(diff_cors))]
print(paste0("Average correlation of sloped and the number of genes: ", mean(diff_cors_selected)))
# Average correlation of CIs and network sizes
diff_cors_selected <- diff_cors[grepl("CI", rownames(diff_cors)), grepl("Number of genes", colnames(diff_cors))]
print(paste0("Average correlation of CIs and the number of genes: ", mean(diff_cors_selected)))

# corrgram::corrgram(diff_cors, order = FALSE, upper.panel=panel.cor)
corrplot.mixed(diff_cors, lower = "number", lower.col = col3(100), upper = "circle", upper.col = col3(100), tl.cex = 1, tl.pos = "lt", tl.col = "black")
# Selected correlations
png(filename = "../manuscript/Figures/FigS1_corr_coherence.png", height = 3000, width = 3000, res = 300)
# corrgram::corrgram(diff_cors, order = FALSE, upper.panel=panel.cor, cex.labels = 4)
corrplot.mixed(diff_cors, lower = "number", lower.col = col3(100), upper = "circle", upper.col = col3(100), tl.cex = 1, tl.pos = "lt", tl.col = "black")
dev.off()
```

\pagebreak

# Summary plots

```{r}
mtx_traits_selected_dotplot <- function(category = "Intelligence", mtx_traits = mtx_traits, databases = c("Biogrid", "STRING full", "STRING filtered")) {
  
  mtx_traits_selected <- mtx_traits[mtx_traits$Category == category, grepl(databases, colnames(mtx_traits))]
  
  mtx_traits_selected_to_plot <- data.frame(Disease = rownames(mtx_traits_selected), 
                                            Pvalue = mtx_traits_selected[, grepl(paste0(databases, " P value"), colnames(mtx_traits_selected)) ], 
                                            Coherence = mtx_traits_selected[, grepl(paste0(databases, " Coherence"), colnames(mtx_traits_selected)) ])
  # Remove noninformative rows
  mtx_traits_selected_to_plot <- mtx_traits_selected_to_plot[!is.na(mtx_traits_selected_to_plot$Coherence), ]
  # Sort by decreasing level of coherence
  mtx_traits_selected_to_plot$Disease <- factor(mtx_traits_selected_to_plot$Disease, levels = mtx_traits_selected_to_plot$Disease[order(mtx_traits_selected_to_plot$Coherence)])
  
  p <- ggplot(data=mtx_traits_selected_to_plot, aes(x=Disease, y=Coherence, colour = "blue", size = Coherence)) + # label = Pvalue, 
    geom_point() +
    theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
    ylab("Coherence") +
    # geom_text_repel(size = 4, colour = "black") +
    coord_flip() +
    # theme_cowplot(12) +
    theme(legend.position="none") +
    theme(axis.text = element_text(size = 12)) + 
    ggtitle(category)
}
```

## Traits

### Fig3_summary_traits.png

```{r results='hide'}
# STRING filtered plots only
p_intelligence <- plot(mtx_traits_selected_dotplot(category = "Intelligence", mtx_traits = mtx0, databases = "String Filt") + theme(axis.title.y = element_blank()))
p_anthropometric <- plot(mtx_traits_selected_dotplot(category = "Anthropometric trait", mtx_traits = mtx0, databases = "String Filt"))
p_cardiovascular <- plot(mtx_traits_selected_dotplot(category = "Cardiovascular trait", mtx_traits = mtx0, databases = "String Filt") + theme(axis.title.y = element_blank()))

p_eye <- plot(mtx_traits_selected_dotplot(category = "Eye trait", mtx_traits = mtx0, databases = "String Filt"))
p_metabolic <- plot(mtx_traits_selected_dotplot(category = "Metabolic trait", mtx_traits = mtx0, databases = "String Filt") + theme(axis.title.y = element_blank()))

png(filename = "../manuscript/Figures/Fig3_summary_traits.png", height = 8, width = 14, units = "in", res = 300)
lay1 <- rbind(c(1, 2, 3))
lay2 <- rbind(c(1, 2, 3))
g1 <- arrangeGrob(grobs = list(p_anthropometric, p_cardiovascular, p_metabolic), layout_matrix = lay1)
g2 <- arrangeGrob(grobs = list(p_eye, p_intelligence), layout_matrix = lay2)
grid.arrange(g1, g2, ncol = 1)
dev.off()
```

### FigS2_summary_traits_Biogrid.png

```{r }
p_intelligence <- plot(mtx_traits_selected_dotplot(category = "Intelligence", mtx_traits = mtx0, databases = "Biogrid") + theme(axis.title.y = element_blank()))
p_anthropometric <- plot(mtx_traits_selected_dotplot(category = "Anthropometric trait", mtx_traits = mtx0, databases = "Biogrid"))
p_cardiovascular <- plot(mtx_traits_selected_dotplot(category = "Cardiovascular trait", mtx_traits = mtx0, databases = "Biogrid") + theme(axis.title.y = element_blank()))
# p_eye <- plot(mtx_traits_selected_dotplot(category = "Eye trait", mtx_traits = mtx0, databases = "Biogrid")) # Insufficient data
p_metabolic <- plot(mtx_traits_selected_dotplot(category = "Metabolic trait", mtx_traits = mtx0, databases = "Biogrid"))

png(filename = "../manuscript/Figures/FigS2_summary_traits_Biogrid.png", height = 6, width = 10, units = "in", res = 300)
lay1 <- rbind(c(1, 2))
lay2 <- rbind(c(1, 2))
g1 <- arrangeGrob(grobs = list(p_anthropometric, p_cardiovascular), layout_matrix = lay1)
g2 <- arrangeGrob(grobs = list(p_metabolic, p_intelligence), layout_matrix = lay2)
grid.arrange(g1, g2, ncol = 1)
dev.off()
```

### FigS3_summary_traits_STRING.png

```{r }
p_intelligence <- plot(mtx_traits_selected_dotplot(category = "Intelligence", mtx_traits = mtx0, databases = "String") + theme(axis.title.y = element_blank()))
p_anthropometric <- plot(mtx_traits_selected_dotplot(category = "Anthropometric trait", mtx_traits = mtx0, databases = "String"))
p_cardiovascular <- plot(mtx_traits_selected_dotplot(category = "Cardiovascular trait", mtx_traits = mtx0, databases = "String") + theme(axis.title.y = element_blank()))
p_eye <- plot(mtx_traits_selected_dotplot(category = "Eye trait", mtx_traits = mtx0, databases = "String"))
p_metabolic <- plot(mtx_traits_selected_dotplot(category = "Metabolic trait", mtx_traits = mtx0, databases = "String") + theme(axis.title.y = element_blank()))

png(filename = "../manuscript/Figures/FigS3_summary_traits_STRING.png", height = 8, width = 14, units = "in", res = 300)
lay1 <- rbind(c(1, 2, 3))
lay2 <- rbind(c(1, 2, 3))
g1 <- arrangeGrob(grobs = list(p_anthropometric, p_cardiovascular, p_metabolic), layout_matrix = lay1)
g2 <- arrangeGrob(grobs = list(p_eye, p_intelligence), layout_matrix = lay2)
grid.arrange(g1, g2, ncol = 1)
dev.off()
```

## Diseases

### Fig4_summary_diseases.png

```{r results='hide'}
# STRING filtered plots only
p_intelligence <- plot(mtx_traits_selected_dotplot(category = "Autoimmune", mtx_traits = mtx0, databases = "String Filt"))
p_anthropometric <- plot(mtx_traits_selected_dotplot(category = "Cancer", mtx_traits = mtx0, databases = "String Filt") + theme(axis.title.y = element_blank())) 
p_cardiovascular <- plot(mtx_traits_selected_dotplot(category = "Cardiovascular disease", mtx_traits = mtx0, databases = "String Filt"))
p_eye <- plot(mtx_traits_selected_dotplot(category = "Eye disease", mtx_traits = mtx0, databases = "String Filt") + theme(axis.title.y = element_blank()))
p_metabolic <- plot(mtx_traits_selected_dotplot(category = "Metabolic disease", mtx_traits = mtx0, databases = "String Filt") + theme(axis.title.y = element_blank()))
p_neurologic <- plot(mtx_traits_selected_dotplot(category = "Neurologic", mtx_traits = mtx0, databases = "String Filt") + theme(axis.title.y = element_blank()))
p_psychiatric <- plot(mtx_traits_selected_dotplot(category = "Psychiatric", mtx_traits = mtx0, databases = "String Filt") + theme(axis.title.y = element_blank())) 

png(filename = "../manuscript/Figures/Fig4_summary_diseases.png", height = 8, width = 14, units = "in", res = 300)
lay1 <- rbind(c(1, 2, 3, 4))
lay2 <- rbind(c(1, 2, 3, 4))
g1 <- arrangeGrob(grobs = list(p_intelligence, p_anthropometric, p_neurologic, p_psychiatric), layout_matrix = lay1)
g2 <- arrangeGrob(grobs = list(p_cardiovascular, p_metabolic, p_eye), layout_matrix = lay2)
grid.arrange(g1, g2, ncol = 1)
dev.off()
```

### FigS4_summary_diseases_Biogrid.png

```{r }
p_intelligence <- plot(mtx_traits_selected_dotplot(category = "Autoimmune", mtx_traits = mtx0, databases = "Biogrid"))
p_anthropometric <- plot(mtx_traits_selected_dotplot(category = "Cancer", mtx_traits = mtx0, databases = "Biogrid") + theme(axis.title.y = element_blank()))
p_cardiovascular <- plot(mtx_traits_selected_dotplot(category = "Cardiovascular disease", mtx_traits = mtx0, databases = "Biogrid"))
p_eye <- plot(mtx_traits_selected_dotplot(category = "Eye disease", mtx_traits = mtx0, databases = "Biogrid") + theme(axis.title.y = element_blank()))
p_metabolic <- plot(mtx_traits_selected_dotplot(category = "Metabolic disease", mtx_traits = mtx0, databases = "Biogrid") + theme(axis.title.y = element_blank()))
# p_neurologic <- plot(mtx_traits_selected_dotplot(category = "Neurologic", mtx_traits = mtx0, databases = "Biogrid") + theme(axis.title.y = element_blank()))
p_psychiatric <- plot(mtx_traits_selected_dotplot(category = "Psychiatric", mtx_traits = mtx0, databases = "Biogrid") + theme(axis.title.y = element_blank()))

png(filename = "../manuscript/Figures/FigS4_summary_diseases_Biogrid.png", height = 8, width = 12, units = "in", res = 300)
lay1 <- rbind(c(1, 2, 3))
lay2 <- rbind(c(1, 2, 3))
g1 <- arrangeGrob(grobs = list(p_intelligence, p_anthropometric, p_psychiatric), layout_matrix = lay1)
g2 <- arrangeGrob(grobs = list(p_cardiovascular, p_metabolic, p_eye), layout_matrix = lay2)
grid.arrange(g1, g2, ncol = 1)
dev.off()
```

### FigS5_summary_diseases_STRING.png

```{r }
p_intelligence <- plot(mtx_traits_selected_dotplot(category = "Autoimmune", mtx_traits = mtx0, databases = "String"))
p_anthropometric <- plot(mtx_traits_selected_dotplot(category = "Cancer", mtx_traits = mtx0, databases = "String") + theme(axis.title.y = element_blank()))
p_cardiovascular <- plot(mtx_traits_selected_dotplot(category = "Cardiovascular disease", mtx_traits = mtx0, databases = "String"))
p_eye <- plot(mtx_traits_selected_dotplot(category = "Eye disease", mtx_traits = mtx0, databases = "String") + theme(axis.title.y = element_blank()))
p_metabolic <- plot(mtx_traits_selected_dotplot(category = "Metabolic disease", mtx_traits = mtx0, databases = "String") + theme(axis.title.y = element_blank()))
p_neurologic <- plot(mtx_traits_selected_dotplot(category = "Neurologic", mtx_traits = mtx0, databases = "String"))
p_psychiatric <- plot(mtx_traits_selected_dotplot(category = "Psychiatric", mtx_traits = mtx0, databases = "String") + theme(axis.title.y = element_blank()))

png(filename = "../manuscript/Figures/FigS5_summary_diseases_STRING.png", height = 12, width = 10, units = "in", res = 300)
lay1 <- rbind(c(1, 2))
lay2 <- rbind(c(1, 2))
lay3 <- rbind(c(1, 2, 3))
g1 <- arrangeGrob(grobs = list(p_intelligence, p_anthropometric), layout_matrix = lay1)
g2 <- arrangeGrob(grobs = list(p_neurologic, p_psychiatric), layout_matrix = lay2)
g3 <- arrangeGrob(grobs = list(p_cardiovascular, p_metabolic, p_eye), layout_matrix = lay3)

grid.arrange(g1, g2, g3, ncol = 1)
dev.off()
```



# Modularity investigation

```{r eval=FALSE}
fileNameIn0  <- "tables/Mod_Table.xlsx"   # diseases and traits, coherence new KEGG, REACTOME, CC. 2019-07-14, Kellen
# iconv -f UTF-8 -t ISO-8859-1 -c Mod_Table.csv >tmp.txt
fileNameIn0  <- "tables/tmp.txt"
mtx0 <- read_csv(fileNameIn0) %>% as.data.frame()
rownames(mtx0) <- mtx0$Disease

# Selected data for correlation
# "Disease", "Category", "Number of SNPs", "Number of genes", "Biogrid Coherence", "String Coherence", "String Filt Coherence", "Biogrid Coherence Reactome", "String Coherence Reactome", "String Filt Coherence Reactome""Biogrid Coherence CC", "String Coherence CC", "String Filt Coherence CC", "Biogrid P-value", "String P-value", "String Filt P-value" 
mtx0_cor <- mtx0 %>% dplyr::select(c("Number of SNPs", "Number of genes", "Biogrid Coherence", "String Coherence", "String Filt Coherence", "Modularity Biogrid", "Modularity String", "Modularity String Filt"))
# Remove large networks
mtx0_cor <- mtx0_cor[mtx0_cor$`Number of genes` < max_genes, ]
# Go through each pairs of columns and calculate pairwise complete observations
# pairwise_columns <- combn(ncol(mtx0_cor), 2)
# for(j in 1:ncol(pairwise_columns)) {
#   print(paste0(colnames(mtx0_cor)[pairwise_columns[1, j]], " - ", colnames(mtx0_cor)[pairwise_columns[2, j]], " : ", nrow(mtx0_cor[complete.cases(mtx0_cor[, c(pairwise_columns[1, j], pairwise_columns[2, j])]), c(pairwise_columns[1, j], pairwise_columns[2, j])])))
# }
# Actual correlation
diff_cors <- cor(mtx0_cor, use = "pairwise.complete.obs")

# corrgram::corrgram(diff_cors, order = FALSE, upper.panel=panel.cor)
corrplot.mixed(diff_cors, lower = "number", lower.col = col3(100), upper = "circle", upper.col = col3(100), tl.cex = 1, tl.pos = "lt", tl.col = "black")
# Selected correlations
png(filename = "Figures/FigS1_corr_coherence.png", height = 2000, width = 2000, res = 300)
# corrgram::corrgram(diff_cors, order = FALSE, upper.panel=panel.cor, cex.labels = 4)
corrplot.mixed(diff_cors, lower = "number", lower.col = col3(100), upper = "circle", upper.col = col3(100), tl.cex = 1, tl.pos = "lt", tl.col = "black")
dev.off()
```


# Confidence Interval investigation

```{r eval=FALSE}
mtx_ci_dotplot <- function(category = "Intelligence", mtx_traits = mtx, databases = c("Biogrid", "String", "String.Filt")) {
  
  mtx_traits_selected <- mtx_traits[mtx_traits$Category == category, grepl(databases, colnames(mtx_traits), fixed = TRUE)]
  
  mtx_traits_selected_to_plot <- data.frame(Disease = rownames(mtx_traits_selected), 
                                            Slope = mtx_traits_selected[, paste0(databases, ".Slope")], 
                                            Lower = mtx_traits_selected[, paste0(databases, ".Lower")],
                                            Upper = mtx_traits_selected[, paste0(databases, ".Upper")])
  # Remove noninformative rows
  mtx_traits_selected_to_plot <- mtx_traits_selected_to_plot[complete.cases(mtx_traits_selected_to_plot), ]
  # Sort by decreasing level of slope
  mtx_traits_selected_to_plot$Disease <- factor(mtx_traits_selected_to_plot$Disease, levels = mtx_traits_selected_to_plot$Disease[order(mtx_traits_selected_to_plot$Slope)])
  
  p <- ggplot() +
    geom_point(data=mtx_traits_selected_to_plot, mapping=aes(x=Disease, y=Slope), size=4, shape=21, fill="white") +
    geom_errorbar(data=mtx_traits_selected_to_plot, mapping=aes(x=Disease, ymin=Lower, ymax=Upper), width=0.2, size=1, color="blue") +  
    theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
    ylab("Slope (unnormalized coherence") +
    # geom_text_repel(size = 4, colour = "black") +
    coord_flip() +
    # theme_cowplot(12) +
    theme(legend.position="none") +
    theme(axis.text = element_text(size = 12)) + 
    ggtitle(category)
}


```
